requirements.txt
# Requirements for Microsoft Fabric CI/CD deployment
# Pin versions for reproducible builds

fabric-cicd==0.1.2
azure-identity==1.19.0
pyyaml==6.0.2
requests>=2.32.0

# Development / CI tooling
ruff>=0.8.0
